use Composer to fetch and load vendor libraries
This patch adds proper Composer usage, in which all the vendor/ libraries are now removed
from this repository and are fetched after cloning/pulling it. This also makes us use the
Composer-generated autoloader, so I removed a lot of manual class-loads.

To clean this up even further, we let Composer always-include the global functions (the
cost of that is virtually zero, but it frees us from having to keep track of what file to include
for what function). This killed a good chunk of other require statements.

JSMin is not maintained anymore, so we should switch to something better soon. CSSMin
is also not Composer-ready, hence the two manual package declarations.

The psr-0 loading for Slim is there because this old (but last 1.x) version of Slim has an error
in its own composer.json, which makes it unloadable.