set the installed_at date when inserting a plugin into the database (fixes #84)
When MySQL is running in the TRADITIONAL mode, trying to insert a row
without a value for a "DATETIME NOT NULL" column leads to an error. By
default, this mode is not set on MySQL/MariaDB.

MySQL 5.6.5+ would support setting a default value for a DATETIME
column, but I don't want to raise the minimum requirement for this sole
reason.