fixed bad path construction
ROOT_PATH is "../" or "../../" most of the time (when not in CLI). The
constructed path would be "../www../charts/static" for the www
environment and look even worse in the api env.

Thankfully, publish_js() is only called when in CLI mode, but it's
broken there as well. ROOT_PATH is absolute and yields something
like "/var/www/website/www/var/www/website/charts/static".

Basically, there is no way this path construction could ever produce
sensible results. ;-)